+++ mobile/android/focus-android/app/src/main/java/org/mozilla/focus/state/AppState.kt
@@ -102,6 +102,7 @@ sealed class Screen {
             Advanced,
             Mozilla,
             About,
+            Licenses,
             Locale,
 
             PrivacyExceptions,